python_backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
selenium
webdriver_manager
twilio
python-dotenv
google-generativeai
